write tests

+++ led_switch_device/src/mesh/security/auth/ModularExp.hpp
@@ -4,7 +4,7 @@
 class ModularExp {
 private:
 public:
-    static uint32_t mod_exp(uint32_t a, uint16_t degree, uint16_t mod) {
+    static uint64_t mod_exp(uint32_t a, uint16_t degree, uint32_t mod) {
         uint32_t c = 1;
         for(int i = 1; i <= degree; i++) {
             c = (c*a)%mod;
